Moved test utilities to separate directory
