+++ Tribler/Test/Core/CreditMining/test_creditmining_sys.py
@@ -23,7 +23,7 @@ from Tribler.Policies.BoostingManager import BoostingManager, BoostingSettings
 from Tribler.Test.Core.CreditMining.mock_creditmining import MockLtTorrent, ResourceFailClass
 from Tribler.Test.common import TORRENT_FILE, TORRENT_FILE_INFOHASH
 from Tribler.Test.test_as_server import TestAsServer, TESTS_DATA_DIR
-from Tribler.Test.util import prepare_xml_rss
+from Tribler.Test.util.util import prepare_xml_rss
 from Tribler.community.allchannel.community import AllChannelCommunity
 from Tribler.community.channel.community import ChannelCommunity
 from Tribler.dispersy.dispersy import Dispersy
