+++ Tribler/Test/util/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains various utilities for the Tribler test suite.
+"""
